update UKKQueue to be 64-bit compatible (although we're still not ready for 64-bits just yet)

+++ macosx/UKKQueue/UKKQueue.m
@@ -269,8 +269,8 @@ static UKKQueue * gUKKQueueSharedQueueSingleton = nil;
 
 -(void) removePathFromQueue: (NSString*)path
 {
-    int		index = 0;
-    int		fd = -1;
+    NSUInteger  index = 0;
+    int         fd = -1;
     
     AT_SYNCHRONIZED( self )
     {
